refactor: rename to StyledBar
