+++ src/components/centerBlock/OverlayPlayer/StyledBar.vue
